Moved ChunkSection info to its own package for speed-optimisation;

+++ octetmc-protocol/src/value/chunk_section_pos.rs
@@ -14,7 +14,7 @@ pub struct ChunkSectionPos {
     /// Y coordinate.
     ///
     /// Positive is up, negative is down.
-    pub y : i32,
+    pub y : u8,
 
     /// Z coordinate.
     ///
